Extract Flutter TTS configuration into a dedicated model file
