+++ sdk/runanywhere-flutter/packages/runanywhere/lib/native/dart_bridge_tts.dart
@@ -9,8 +9,7 @@ import 'dart:isolate';
 import 'dart:typed_data';
 
 import 'package:ffi/ffi.dart';
-import 'package:runanywhere/features/tts/system_tts_service.dart'
-    show TTSConfiguration;
+import 'package:runanywhere/features/tts/tts_configuration.dart';
 import 'package:runanywhere/foundation/logging/sdk_logger.dart';
 import 'package:runanywhere/native/ffi_types.dart';
 import 'package:runanywhere/native/platform_loader.dart';
