+++ sdk/runanywhere-flutter/packages/runanywhere/lib/runanywhere.dart
@@ -14,7 +14,7 @@ export 'core/types/storage_types.dart';
 export 'data/network/network.dart';
 export 'features/llm/llm_configuration.dart';
 export 'features/stt/stt_configuration.dart';
-export 'features/tts/system_tts_service.dart' show TTSConfiguration;
+export 'features/tts/tts_configuration.dart';
 export 'features/vad/vad_configuration.dart';
 export 'foundation/configuration/sdk_constants.dart';
 export 'foundation/error_types/sdk_error.dart';
